Moved RelationshipMetadata logic to a new service
