+++ dspace-api/src/main/java/org/dspace/content/service/ItemService.java
@@ -655,19 +655,6 @@ public interface ItemService
     boolean isInProgressSubmission(Context context, Item item) throws SQLException;
 
     /**
-     * This method retrieves a list of MetadataValue objects that get constructed from processing
-     * the given Item's Relationships through the config given to the {@link VirtualMetadataPopulator}
-     * @param item  The Item that will be processed through it's Relationships
-     * @param enableVirtualMetadata This parameter will determine whether the list of Relationship metadata
-     *                              should be populated with metadata that is being generated through the
-     *                              VirtualMetadataPopulator functionality or not
-     * @return      The list of MetadataValue objects constructed through the Relationships
-     */
-    public List<RelationshipMetadataValue> getRelationshipMetadata(Item item, boolean enableVirtualMetadata);
-
-
-
-    /**
      * Get metadata for the DSpace Object in a chosen schema.
      * See <code>MetadataSchema</code> for more information about schemas.
      * Passing in a <code>null</code> value for <code>qualifier</code>
