+++ dspace-api/src/test/java/org/dspace/content/RelationshipMetadataServiceTest.java
@@ -1,0 +1,54 @@
+package org.dspace.content;
+
+import static org.mockito.Mockito.when;
+
+import java.util.LinkedList;
+
+import org.dspace.content.service.RelationshipService;
+import org.dspace.core.Context;
+import org.junit.Before;
+import org.junit.Ignore;
+import org.junit.Test;
+import org.junit.runner.RunWith;
+import org.mockito.InjectMocks;
+import org.mockito.Mock;
+import org.mockito.runners.MockitoJUnitRunner;
+
+@RunWith(MockitoJUnitRunner.class)
+public class RelationshipMetadataServiceTest {
+
+    @InjectMocks
+    private RelationshipMetadataServiceImpl relationshipMetadataService;
+
+    @Mock
+    private RelationshipService relationshipService;
+
+    @Mock
+    private Item item;
+    private Relationship firstRelationship;
+    private Relationship secondRelationship;
+    private LinkedList<Relationship> list;
+    private Context context;
+
+    @Before
+    public void init() {
+        firstRelationship = new Relationship();
+        secondRelationship = new Relationship();
+        list = new LinkedList<>();
+        list.add(firstRelationship);
+        list.add(secondRelationship);
+    }
+
+    @Test
+    @Ignore
+    public void testGetRelationshipMetadata() throws Exception {
+        when(item.getMetadata()).thenReturn(new LinkedList<>());
+        when(relationshipService.findByItem(context, item)).thenReturn(list);
+
+
+        String t = "";
+        relationshipMetadataService.getRelationshipMetadata(item, true);
+
+
+    }
+}
